pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mobilenetv2-ssd"
version = "0.1.0"
description = "MobileNetV2-SSD object detection pipeline"
requires-python = ">=3.10"

dependencies = [
  "tensorflow==2.17.0",
  "numpy>=1.26.0",
  "Pillow>=10.0.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-cov"
]

onnx-export = [
  "onnxruntime>=1.17.0",
  "tf2onnx>=1.16.0",
  "onnx>=1.15.0",
]

cloud = [
  "boto3>=1.34.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -m 'not integration and not regression'"

markers = [
  "unit: fast, deterministic unit tests",
  "integration: slower tests that touch multiple components",
  "regression: tests that lock down previously-broken behavior"
]
